Refactor settings persistence using modular Preferences DataStore (Option B)
