+++ core/camera/testing/build.gradle.kts
@@ -52,6 +52,7 @@ dependencies {
     implementation(project(":core:camera"))
     implementation(project(":core:model"))
     implementation(project(":data:settings"))
+    implementation(project(":core:settings"))
     
     implementation(libs.camera.core)
     implementation(libs.kotlinx.coroutines.core)
