+++ core/settings/build.gradle.kts
@@ -17,10 +17,12 @@
 plugins {
     alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
+    alias(libs.plugins.kotlin.kapt)
+    alias(libs.plugins.dagger.hilt.android)
 }
 
 android {
-    namespace = "com.google.jetpackcamera.settings.testing"
+    namespace = "com.google.jetpackcamera.core.settings"
     compileSdk = libs.versions.compileSdk.get().toInt()
 
     defaultConfig {
@@ -29,14 +31,7 @@ android {
         lint.targetSdk = libs.versions.targetSdk.get().toInt()
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-    }
-
-    flavorDimensions += "flavor"
-    productFlavors {
-        create("stable") {
-            dimension = "flavor"
-            isDefault = true
-        }
+        consumerProguardFiles("consumer-rules.pro")
     }
 
     compileOptions {
@@ -49,9 +44,16 @@ android {
 }
 
 dependencies {
-    implementation(project(":data:settings"))
-    implementation(project(":core:settings"))
-    implementation(project(":core:model"))
-    
     implementation(libs.kotlinx.coroutines.core)
+
+    // Hilt
+    implementation(libs.dagger.hilt.android)
+    kapt(libs.dagger.hilt.compiler)
+
+    // Domain models
+    implementation(project(":core:model"))
+}
+
+kapt {
+    correctErrorTypes = true
 }
