+++ core/settings/consumer-rules.pro
@@ -1,0 +1,1 @@
+# Consumer proguard rules for core:settings
