+++ core/settings/datastore-prefs/.gitignore
