+++ core/settings/datastore-prefs/consumer-rules.pro
