+++ core/settings/datastore-prefs/src/main/AndroidManifest.xml
