+++ core/settings/src/main/AndroidManifest.xml
