+++ core/settings/src/main/java/com/google/jetpackcamera/settings/model/Constraints.kt
