+++ data/settings/src/main/java/com/google/jetpackcamera/settings/LocalSettingsRepository.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.jetpackcamera.settings.testing
+package com.google.jetpackcamera.settings
 
 import com.google.jetpackcamera.model.AspectRatio
 import com.google.jetpackcamera.model.DarkMode
@@ -25,70 +25,72 @@ import com.google.jetpackcamera.model.LowLightBoostPriority
 import com.google.jetpackcamera.model.StabilizationMode
 import com.google.jetpackcamera.model.StreamConfig
 import com.google.jetpackcamera.model.VideoQuality
-import com.google.jetpackcamera.settings.SettingsRepository
 import com.google.jetpackcamera.settings.model.CameraAppSettings
-import com.google.jetpackcamera.settings.model.DEFAULT_CAMERA_APP_SETTINGS
+import javax.inject.Inject
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
 
-class FakeSettingsRepository(
-    initialSettings: CameraAppSettings = DEFAULT_CAMERA_APP_SETTINGS
+/**
+ * Implementation of [SettingsRepository] delegating to [SettingsDataSource].
+ */
+class LocalSettingsRepository @Inject constructor(
+    private val settingsDataSource: SettingsDataSource
 ) : SettingsRepository {
-    private val _defaultCameraAppSettings = MutableStateFlow(initialSettings)
-    override val defaultCameraAppSettings: Flow<CameraAppSettings> = _defaultCameraAppSettings.asStateFlow()
 
-    override suspend fun getCurrentDefaultCameraAppSettings() = _defaultCameraAppSettings.value
+    override val defaultCameraAppSettings: Flow<CameraAppSettings> =
+        settingsDataSource.defaultCameraAppSettings
+
+    override suspend fun getCurrentDefaultCameraAppSettings(): CameraAppSettings =
+        settingsDataSource.getCurrentDefaultCameraAppSettings()
 
     override suspend fun updateDefaultLensFacing(lensFacing: LensFacing) {
-        _defaultCameraAppSettings.value = _defaultCameraAppSettings.value.copy(cameraLensFacing = lensFacing)
+        settingsDataSource.updateDefaultLensFacing(lensFacing)
     }
 
     override suspend fun updateDarkModeStatus(darkMode: DarkMode) {
-        _defaultCameraAppSettings.value = _defaultCameraAppSettings.value.copy(darkMode = darkMode)
+        settingsDataSource.updateDarkModeStatus(darkMode)
     }
 
     override suspend fun updateFlashModeStatus(flashMode: FlashMode) {
-        _defaultCameraAppSettings.value = _defaultCameraAppSettings.value.copy(flashMode = flashMode)
-    }
-
-    override suspend fun updateStreamConfig(streamConfig: StreamConfig) {
-        _defaultCameraAppSettings.value = _defaultCameraAppSettings.value.copy(streamConfig = streamConfig)
-    }
-
-    override suspend fun updateLowLightBoostPriority(lowLightBoostPriority: LowLightBoostPriority) {
-        _defaultCameraAppSettings.value = _defaultCameraAppSettings.value.copy(lowLightBoostPriority = lowLightBoostPriority)
-    }
-
-    override suspend fun updateStabilizationMode(stabilizationMode: StabilizationMode) {
-        _defaultCameraAppSettings.value = _defaultCameraAppSettings.value.copy(stabilizationMode = stabilizationMode)
-    }
-
-    override suspend fun updateDynamicRange(dynamicRange: DynamicRange) {
-        _defaultCameraAppSettings.value = _defaultCameraAppSettings.value.copy(dynamicRange = dynamicRange)
+        settingsDataSource.updateFlashModeStatus(flashMode)
     }
 
     override suspend fun updateAspectRatio(aspectRatio: AspectRatio) {
-        _defaultCameraAppSettings.value = _defaultCameraAppSettings.value.copy(aspectRatio = aspectRatio)
+        settingsDataSource.updateAspectRatio(aspectRatio)
+    }
+
+    override suspend fun updateStreamConfig(streamConfig: StreamConfig) {
+        settingsDataSource.updateStreamConfig(streamConfig)
+    }
+
+    override suspend fun updateLowLightBoostPriority(lowLightBoostPriority: LowLightBoostPriority) {
+        settingsDataSource.updateLowLightBoostPriority(lowLightBoostPriority)
+    }
+
+    override suspend fun updateStabilizationMode(stabilizationMode: StabilizationMode) {
+        settingsDataSource.updateStabilizationMode(stabilizationMode)
+    }
+
+    override suspend fun updateDynamicRange(dynamicRange: DynamicRange) {
+        settingsDataSource.updateDynamicRange(dynamicRange)
     }
 
     override suspend fun updateTargetFrameRate(targetFrameRate: Int) {
-        _defaultCameraAppSettings.value = _defaultCameraAppSettings.value.copy(targetFrameRate = targetFrameRate)
+        settingsDataSource.updateTargetFrameRate(targetFrameRate)
     }
 
     override suspend fun updateImageFormat(imageFormat: ImageOutputFormat) {
-        _defaultCameraAppSettings.value = _defaultCameraAppSettings.value.copy(imageFormat = imageFormat)
+        settingsDataSource.updateImageFormat(imageFormat)
     }
 
     override suspend fun updateMaxVideoDuration(durationMillis: Long) {
-        _defaultCameraAppSettings.value = _defaultCameraAppSettings.value.copy(maxVideoDurationMillis = durationMillis)
+        settingsDataSource.updateMaxVideoDuration(durationMillis)
     }
 
     override suspend fun updateVideoQuality(videoQuality: VideoQuality) {
-        _defaultCameraAppSettings.value = _defaultCameraAppSettings.value.copy(videoQuality = videoQuality)
+        settingsDataSource.updateVideoQuality(videoQuality)
     }
 
     override suspend fun updateAudioEnabled(isAudioEnabled: Boolean) {
-        _defaultCameraAppSettings.value = _defaultCameraAppSettings.value.copy(audioEnabled = isAudioEnabled)
+        settingsDataSource.updateAudioEnabled(isAudioEnabled)
     }
 }
