+++ settings.gradle.kts
@@ -57,6 +57,7 @@ include(":ui:uistateadapter:capture")
 include(":ui:components")
 include(":ui:components:capture")
 include(":data:model")
+include(":core:settings")
 include(":core:model")
 include(":ui:uistate:postcapture")
 include(":ui:uistateadapter:postcapture")
@@ -64,5 +65,5 @@ include(":core:camera:postprocess")
 include(":ui:controller")
 include(":ui:controller:impl")
 include(":ui:controller:testing")
-include(":data:settings-datastore")
-include(":data:settings-datastore:testing")
+include(":core:settings:datastore-prefs")
+include(":core:settings:datastore-prefs:testing")
