+++ ui/components/capture/build.gradle.kts
@@ -108,6 +108,7 @@ dependencies {
     testImplementation(project(":core:common"))
     testImplementation(project(":core:camera:testing"))
     testImplementation(project(":data:settings"))
+    testImplementation(project(":core:settings"))
 
 }
